Feat : 자유게시판 피드부분 모달 기능 코드를 피드카드로 변경
